+++ lib/src/widgets/embeds/image_resizer.dart
@@ -3,8 +3,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 
-import '../../translations/toolbar.i18n.dart';
-
 class ImageResizer extends StatefulWidget {
   const ImageResizer(
       {required this.imageWidth,
@@ -77,7 +75,7 @@ class _ImageResizerState extends State<ImageResizer> {
             value: value,
             max: max,
             divisions: 1000,
-            label: label.i18n,
+            label: label,
             onChanged: (val) {
               setState(() {
                 onChanged(val);