+++ lib/src/widgets/link.dart
@@ -4,7 +4,6 @@ import 'package:flutter/material.dart';
 
 import '../models/documents/attribute.dart';
 import '../models/documents/nodes/node.dart';
-import '../translations/toolbar.i18n.dart';
 
 const linkPrefixes = [
   'mailto:', // email
@@ -166,17 +165,17 @@ Future<LinkMenuAction> _showMaterialMenu(
         mainAxisSize: MainAxisSize.min,
         children: [
           _MaterialAction(
-            title: 'Open'.i18n,
+            title: 'Open',
             icon: Icons.language_sharp,
             onPressed: () => Navigator.of(context).pop(LinkMenuAction.launch),
           ),
           _MaterialAction(
-            title: 'Copy'.i18n,
+            title: 'Copy',
             icon: Icons.copy_sharp,
             onPressed: () => Navigator.of(context).pop(LinkMenuAction.copy),
           ),
           _MaterialAction(
-            title: 'Remove'.i18n,
+            title: 'Remove',
             icon: Icons.link_off_sharp,
             onPressed: () => Navigator.of(context).pop(LinkMenuAction.remove),
           ),