+++ lib/src/widgets/toolbar/image_video_utils.dart
@@ -5,7 +5,6 @@ import 'package:image_picker/image_picker.dart';
 import '../../models/documents/nodes/embeddable.dart';
 import '../../models/rules/insert.dart';
 import '../../models/themes/quill_dialog_theme.dart';
-import '../../translations/toolbar.i18n.dart';
 import '../../utils/platform.dart';
 import '../controller.dart';
 import '../toolbar.dart';
@@ -40,7 +39,7 @@ class LinkDialogState extends State<LinkDialog> {
         maxLines: null,
         style: widget.dialogTheme?.inputTextStyle,
         decoration: InputDecoration(
-            labelText: 'Paste a link'.i18n,
+            labelText: 'Paste a link',
             labelStyle: widget.dialogTheme?.labelTextStyle,
             floatingLabelStyle: widget.dialogTheme?.labelTextStyle),
         autofocus: true,
@@ -54,7 +53,7 @@ class LinkDialogState extends State<LinkDialog> {
               ? _applyLink
               : null,
           child: Text(
-            'Ok'.i18n,
+            'Ok',
             style: widget.dialogTheme?.labelTextStyle,
           ),
         ),
@@ -95,7 +94,7 @@ class ImageVideoUtils {
                   Icons.collections,
                   color: Colors.orangeAccent,
                 ),
-                label: Text('Gallery'.i18n),
+                label: Text('Gallery'),
                 onPressed: () => Navigator.pop(ctx, MediaPickSetting.Gallery),
               ),
               TextButton.icon(
@@ -103,7 +102,7 @@ class ImageVideoUtils {
                   Icons.link,
                   color: Colors.cyanAccent,
                 ),
-                label: Text('Link'.i18n),
+                label: Text('Link'),
                 onPressed: () => Navigator.pop(ctx, MediaPickSetting.Link),
               )
             ],